Separated history package
