+++ app/src/main/java/com/example/runtracker/MainActivity.kt
@@ -14,6 +14,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.runtracker.gallery.GalleryActivity
+import com.example.runtracker.history.HistoryActivity
 import com.example.runtracker.myAccount.MyAccountActivity
 import com.example.runtracker.runRecording.MapActivity
 import com.example.runtracker.settings.SettingsActivity
