+++ app/src/main/java/com/example/runtracker/history/HistoryActivity.kt
@@ -1,4 +1,4 @@
-package com.example.runtracker
+package com.example.runtracker.history
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,6 +6,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.runtracker.*
 import com.example.runtracker.database.Run
 
 class HistoryActivity : AppCompatActivity(), RunHistoryAdapter.OnRunItemClickListener {
